Refactor: Manually load file into editor
Previously we have an effect that will load a file's content into the editor when the handle is changed. This turns out to be difficult to handle the "new" and "save as" cases.

This PR removes the effect and use a more imperative approach, in which the action to load file content is handled manually in relevant handlers. For example, the handler that open a new file will also call "openFileInEditor" to update the editor right there.

Thank to the new approach, implementing "New" and "Save As" is now straightforward, and is also done in this PR